Module jortestkit::console::style::icons

icons are not meant to add some SWAG to the display, not to use as a reference items. When using Icons, remember some users don’t have unicode compatible/enabled terminals so always use this to decorate.

Structs

account

block

blockchain

failure

ignored

jcli

icon associated the jcli command line binary

jormungandr

icon associated the jcli command line binary

scenario

seed

icon associated the test Seed to re-run the same tests with the same cryptographic materials.

success

transaction

wallet

Type Definitions

Icon